#182 add audio exists validaion
